Hack to avoid DRC errors
Now that we've using the "unwrapped" GPIO cell, we have to route to two
close pins on m3. When the cell is vertical (opposite of the m3
preferred direction), the router sometimes does weird things that result
in m3 shapes too close together. Extending the two pins away from each
other in opposite directions seems to solve this problem.