Small chip DRC/LVS clean!
Needed to go back to wrapped GPIO cells, since it seemed like the
OpenROAD router was having a really hard time not creating DRC
violations when routing to the less uniform pins on the foundry cells.

In order to support this, we also had to add extra bits of metal on the
padring side to connect the power pads to the wiring. I did it this way
since we needed to have a couple power pins for LVS, but OpenROAD
doesn't seem to like pins that extend beyond the core area.