Some changes, no more routing congestion
Main thing: increased diesize by 1.2^2 and placed north and east pads
along the first half of each side. Turns out the routing congestion was
actually resulting in violations (that were not being caught by
top-level DRC). It seems like the congestion is related to routing to
I/O around the SRAM: the DRC errors were wires that were overlapping
with the SRRAM boundary.

I also added the tool option to set placement density to 0.15 and added
a call to insert_vias().